Move user agent substring generation from `Bridge` to `Strada`.
