+++ Source/Bridge.swift
@@ -29,11 +29,6 @@ public final class Bridge: Bridgable {
         }
     }
     
-    public static func userAgentSubstring(for componentTypes: [BridgeComponent.Type]) -> String {
-        let components = componentTypes.map { $0.name }.joined(separator: " ")
-        return "bridge-components: [\(components)]"
-    }
-    
     init(webView: WKWebView) {
         self.webView = webView
         loadIntoWebView()
